functions/risk_evaluator/requirements.txt
functions-framework==3.5.0
google-cloud-firestore==2.18.0
google-cloud-pubsub==2.23.0
